refactor(hooks): extract useAttendance and useCurriculum custom hooks (#1546)
Eliminates duplicated data-fetching logic across Student, Teacher,
and Institute dashboards by extracting shared custom hooks.

New files:
- hooks/useAttendance.js — shared attendance fetching hook
  Accepts { role, id, options }, returns { attendance, loading, error, refetch }
  Handles role-specific endpoints via switch on role param
  useCallback memoizes fetch to prevent useEffect infinite loops

- hooks/useCurriculum.js — shared curriculum fetching hook
  Accepts { role, id }, returns { curriculum, loading, error, refetch }
  Same pattern and error handling as useAttendance

Refactored:
- components/StudentDashboard.js — removed inline attendance/curriculum
  useState + useEffect + fetch blocks, replaced with hook calls
- components/TeacherDashboardComponent.js — same refactor for teacher role
- components/InstituteDashboard.js — same refactor for institute role

No functional changes — loading states, error states, and rendered data
are identical. All component JSX untouched. Only data-fetching logic moved.
refetch() enables manual re-fetch without page reload.

Zero new dependencies. npm run build passes.

Closes #1546
