+++ app/dashboard/loading.jsx
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import Skeleton from "@/components/ui/Skeleton";
