+++ app/student/dashboard/loading.jsx
@@ -1,3 +1,4 @@
+"use client";
 import DashboardSkeleton from "@/components/ui/DashboardSkeleton";
 
 export default function Loading() {
